refac: create component for ui detail
